Update to correct WITH rather than USE
Updating as USE is old and WITH is the only way to get it to compile successfully if you have compiled ADCore with GraphicsMagick. Also modified the findstring method of confirming the target architecture as it could spit out multiple variations of windows-x64 or wind32-x86 such as -static, -mingw for instance.